+++ src/ExpoYandexMapKitView.web.tsx
@@ -24,6 +24,7 @@ export const YandexMapView = React.forwardRef<YandexMapViewRef, YandexMapViewPro
         selectGeoObject: unsupported,
         deselectGeoObject: unsupported,
         fitToCoordinates: unsupported,
+        setIndoorLevel: unsupported,
         addTileOverlay: unsupported,
         removeTileOverlay: unsupported,
       };